Major update
I refactored the code a bit, there is a lot of work to be done but hey, it's something.

I also improve on tictactoe.java and added the capability for 2 people to play the game as well.

I also did some polishing but like I said there is still a lot of work to be done.